+++ examples/spot/account/getAccountCurrencies.js
@@ -1,13 +1,15 @@
 'use strict'
 
+
 const BitMart = require('../../../src')
 const { logger } = require('../../config')
 const bitmartSpotAPI = new BitMart.BitmartSpotAPI({
     logger: logger,
 })
 
-bitmartSpotAPI.getCurrencies()
+
+bitmartSpotAPI.getAccountCurrencies({
+    currencies: 'BTC,ETH,USDT'
+})
   .then(response => bitmartSpotAPI.logger.log(response.data))
   .catch(error => bitmartSpotAPI.logger.log(error))
-
-