+++ examples/spot/account/getDepositAndWithdrawHistory.js
@@ -1,15 +1,20 @@
 'use strict'
 
-const BitmartSpotAPI = require('../../../src/bitmartSpotAPI')
-
-const yourApiKey = 'your api key'
-const bitmartSpotAPI = new BitmartSpotAPI({
-  apiKey: yourApiKey
+const BitMart = require('../../../src')
+const { yourApiKey, logger } = require('../../config')
+const bitmartSpotAPI = new BitMart.BitmartSpotAPI({
+  apiKey: yourApiKey,
+  logger: logger,
 })
 
 
+const endTime = Date.now()
+const startTime = endTime - 24 * 60 * 60 * 1000 // a day ago
+
 bitmartSpotAPI.getDepositAndWithdrawHistory('withdraw', 10, {
-    currency: "BTC"
+    currency: "BTC",
+    startTime: startTime,
+    endTime: endTime
 })
   .then(response => bitmartSpotAPI.logger.log(response.data))
   .catch(error => bitmartSpotAPI.logger.log(error))