+++ examples/spot/account/getMarginAccountDetails.js
@@ -1,10 +1,10 @@
 'use strict'
 
-const BitmartSpotAPI = require('../../../src/bitmartSpotAPI')
-
-const yourApiKey = 'your api key'
-const bitmartSpotAPI = new BitmartSpotAPI({
-  apiKey: yourApiKey
+const BitMart = require('../../../src')
+const { yourApiKey, logger } = require('../../config')
+const bitmartSpotAPI = new BitMart.BitmartSpotAPI({
+  apiKey: yourApiKey,
+  logger: logger,
 })