+++ examples/spot/account/getSpotAccountBalance.js
@@ -1,19 +1,29 @@
 'use strict'
 
-const BitmartSpotAPI = require('../../../src/bitmartSpotAPI')
-
-const yourApiKey = 'your api key'
-const bitmartSpotAPI = new BitmartSpotAPI({
-  apiKey: yourApiKey
+const BitMart = require('../../../src')
+const { yourApiKey, logger } = require('../../config')
+const bitmartSpotAPI = new BitMart.BitmartSpotAPI({
+  apiKey: yourApiKey,
+  logger: logger,
 })
 
 bitmartSpotAPI.getSpotAccountBalance({
-  currency: 'USDT'
+  currency: 'USDT',
+  needUsdValuation: true
 })
   .then(response => bitmartSpotAPI.logger.log(response.data))
   .catch(error => bitmartSpotAPI.logger.log(error))
+
+
+bitmartSpotAPI.getSpotAccountBalance({
+  needUsdValuation: false
+})
+  .then(response => bitmartSpotAPI.logger.log(response.data))
+  .catch(error => bitmartSpotAPI.logger.log(error))
+
 
 
 bitmartSpotAPI.getSpotAccountBalance()
   .then(response => bitmartSpotAPI.logger.log(response.data))
   .catch(error => bitmartSpotAPI.logger.log(error))
+